+++ crates/hotshot-state-prover/Cargo.toml
@@ -28,6 +28,9 @@ serde = { workspace = true, features = ["rc"] }
 tagged-base64 = { git = "https://github.com/espressosystems/tagged-base64", tag = "0.3.0" }
 typenum = { workspace = true }
 
+[dev-dependencies]
+hotshot-stake-table = { path = "../hotshot-stake-table" }
+
 [features]
 default = ["parallel"]
 std = ["ark-std/std", "ark-serialize/std", "ark-ff/std"]
